requirements.txt
# MVP核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-okx==0.4.0
requests==2.31.0
aiofiles==23.2.1
pandas==2.1.3
numpy==1.26.0
